refactor(rlb): move rlb function to dedicated file
