+++ tests/index.js
@@ -1,4 +1,5 @@
 import rlb from '../src/index.js';
+import trueRlb from '../src/rlb.js';
 import * as rlbPackage from '../src/index.js';
 
 describe( 'package', () => {
@@ -10,6 +11,10 @@ describe( 'package', () => {
 	describe( 'default export', () => {
 		it( 'is a function', () => {
 			expect( rlb ).to.be.a( 'function' );
+		} );
+
+		it( 'is the same as rlb from rlb.js file', () => {
+			expect( rlb ).to.equal( trueRlb );
 		} );
 	} );
 } );
